requirements.txt
pandas
numpy
faiss-cpu
sentence-transformers
langchain
langchain-openai
langchain-core
pdfplumber
openpyxl
tabulate
python-dotenv
